+++ CHANGELOG.md
@@ -30,6 +30,9 @@
 * Optimize `allbut.page`. Push the page down the tree if allbut is
   key preserving.
 
+* Optimize `allbut.matching`. Push the matching down the tree, it's
+  always possible.
+
 # 0.18.0 - 2021/03/12
 
 * Default Relation#type is provided, that returns Bmg::Type::ANY